minirt empiueza a ir

+++ miniRT/v1/shapes/libshapes.h
@@ -19,7 +19,7 @@ typedef	struct		s_triangle
 
 t_sphere	sphere(t_point po, float r);
 t_sphere*	new_sphere(t_point po, float r);
-float		intersect_sphere(t_sphere *sp, t_ray ray);
+double		intersect_sphere(t_sphere *sp, t_ray ray);
 t_vector	normal_sphere(t_sphere *sp, t_point po);
 void		print_sphere(t_sphere *sp);
